[HWL8012] Fix stability issues on ESP8266 with high current load
The CF and CF1 pulses are correlated with energy consumption.
So at high load, the ESP8266 had issues keeping up with the 1000's of ISR calls.

Added some caching so we don't need to do as many floating point operations and access volatile variables.
Also the ISR functions are more simplified.